testbench: fix signed/unsigned comparison warning

+++ source/test/ipfilterharness.cpp
@@ -66,7 +66,7 @@ IPFilterHarness::IPFilterHarness()
     }
 
     /*[0] --- Random values  */
-    for (int i = 0; i < ipf_t_size * sizeof(pixel); i++)
+    for (size_t i = 0; i < ipf_t_size * sizeof(pixel); i++)
         pixel_test_buff[0][i] = rand() & PIXEL_MAX;
 
     /*[1] --- Minimum       */
